helloworld: Move HelloWorld class into seperated file
